pass runtime in click context instead of global; move project commands to separate module
