+++ cumulusci/cli/tests/test_cci.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from cumulusci.core.dependencies.dependencies import PackageNamespaceVersionDependency
 from datetime import date
 from datetime import datetime
 from datetime import timedelta
@@ -29,7 +28,6 @@ from cumulusci.core.config import ScratchOrgConfig
 from cumulusci.core.config import BaseProjectConfig
 from cumulusci.core.flowrunner import FlowCoordinator
 from cumulusci.core.exceptions import FlowNotFoundError
-from cumulusci.core.exceptions import NotInProject
 from cumulusci.core.exceptions import OrgNotFound
 from cumulusci.core.exceptions import ScratchOrgException
 from cumulusci.core.exceptions import ServiceNotConfigured
@@ -37,7 +35,7 @@ from cumulusci.core.exceptions import CumulusCIException
 from cumulusci.cli import cci
 from cumulusci.cli.runtime import CliRuntime
 from cumulusci.utils import temporary_dir
-from cumulusci.cli.tests.utils import run_click_command, recursive_list_files, DummyTask
+from cumulusci.cli.tests.utils import run_click_command, DummyTask
 
 MagicMock = mock.MagicMock()
 
@@ -124,23 +122,6 @@ class TestCCI(unittest.TestCase):
         cci.check_latest_version()
 
         click.echo.assert_any_call("Error checking cci version:", err=True)
-
-    def test_render_recursive(self):
-        out = []
-        with mock.patch("click.echo", out.append):
-            cci.render_recursive(
-                {"test": [{"list": ["list"], "dict": {"key": "value"}, "str": "str"}]}
-            )
-        self.assertEqual(
-            """\x1b[1mtest:\x1b[0m
-    -
-        \x1b[1mlist:\x1b[0m
-            - list
-        \x1b[1mdict:\x1b[0m
-            \x1b[1mkey:\x1b[0m value
-        \x1b[1mstr:\x1b[0m str""",
-            "\n".join(out),
-        )
 
     @mock.patch("cumulusci.cli.cci.tee_stdout_stderr")
     @mock.patch("cumulusci.cli.cci.get_tempfile_logger")
@@ -683,219 +664,6 @@ Environment Info: Rossian / x68_46
         run_click_command(cci.service_connect)
         # no assertion; this test is for coverage of empty methods
 
-    def test_validate_project_name(self):
-        with self.assertRaises(click.UsageError):
-            cci.validate_project_name("with spaces")
-
-    def test_validate_project_name__valid(self):
-        assert cci.validate_project_name("valid") == "valid"
-
-    @mock.patch("cumulusci.cli.cci.click")
-    def test_project_init(self, click):
-        with temporary_dir():
-            os.mkdir(".git")
-            Path(".git", "HEAD").write_text("ref: refs/heads/main")
-
-            click.prompt.side_effect = (
-                "testproj",  # project_name
-                "testpkg",  # package_name
-                "testns",  # package_namespace
-                "43.0",  # api_version
-                "mdapi",  # source_format
-                "3",  # extend other URL
-                "https://github.com/SalesforceFoundation/NPSP",  # github_url
-                "main",  # git_default_branch
-                "work/",  # git_prefix_feature
-                "uat/",  # git_prefix_beta
-                "rel/",  # git_prefix_release
-                "%_TEST%",  # test_name_match
-                "90",  # code_coverage
-            )
-            click.confirm.side_effect = (
-                True,
-                True,
-                True,
-            )  # is managed? extending? enforce Apex coverage?
-
-            runtime = CliRuntime(
-                config={"project": {"test": {"name_match": "%_TEST%"}}},
-                load_keychain=False,
-            )
-            run_click_command(cci.project_init, runtime=runtime)
-
-            # Make sure expected files/dirs were created
-            self.assertEqual(
-                [
-                    ".git/",
-                    ".git/HEAD",
-                    ".github/",
-                    ".github/PULL_REQUEST_TEMPLATE.md",
-                    ".gitignore",
-                    "README.md",
-                    "cumulusci.yml",
-                    "datasets/",
-                    "datasets/mapping.yml",
-                    "orgs/",
-                    "orgs/beta.json",
-                    "orgs/dev.json",
-                    "orgs/feature.json",
-                    "orgs/release.json",
-                    "robot/",
-                    "robot/testproj/",
-                    "robot/testproj/doc/",
-                    "robot/testproj/resources/",
-                    "robot/testproj/tests/",
-                    "robot/testproj/tests/create_contact.robot",
-                    "sfdx-project.json",
-                    "src/",
-                ],
-                recursive_list_files(),
-            )
-
-    @mock.patch("cumulusci.cli.cci.click")
-    def test_project_init_tasks(self, click):
-        """Verify that the generated cumulusci.yml file is readable and has the proper robot task"""
-        with temporary_dir():
-            os.mkdir(".git")
-            Path(".git", "HEAD").write_text("ref: refs/heads/main")
-
-            click.prompt.side_effect = (
-                "testproj",  # project_name
-                "testpkg",  # package_name
-                "testns",  # package_namespace
-                "43.0",  # api_version
-                "mdapi",  # source_format
-                "3",  # extend other URL
-                "https://github.com/SalesforceFoundation/NPSP",  # github_url
-                "main",  # git_default_branch
-                "work/",  # git_prefix_feature
-                "uat/",  # git_prefix_beta
-                "rel/",  # git_prefix_release
-                "%_TEST%",  # test_name_match
-                "90",  # code_coverage
-            )
-            click.confirm.side_effect = (
-                True,
-                True,
-                True,
-            )  # is managed? extending? enforce code coverage?
-
-            run_click_command(cci.project_init)
-
-            # verify we can load the generated yml
-            cli_runtime = CliRuntime(load_keychain=False)
-
-            # ...and verify it has the expected tasks
-            config = cli_runtime.project_config.config_project
-            expected_tasks = {
-                "robot": {
-                    "options": {
-                        "suites": "robot/testproj/tests",
-                        "options": {"outputdir": "robot/testproj/results"},
-                    }
-                },
-                "robot_testdoc": {
-                    "options": {
-                        "path": "robot/testproj/tests",
-                        "output": "robot/testproj/doc/testproj_tests.html",
-                    }
-                },
-                "run_tests": {"options": {"required_org_code_coverage_percent": 90}},
-            }
-            self.assertDictEqual(config["tasks"], expected_tasks)
-
-    def test_project_init_no_git(self):
-        with temporary_dir():
-            with self.assertRaises(click.ClickException):
-                run_click_command(cci.project_init)
-
-    def test_project_init_already_initted(self):
-        with temporary_dir():
-            os.mkdir(".git")
-            Path(".git", "HEAD").write_text("ref: refs/heads/main")
-            with open("cumulusci.yml", "w"):
-                pass  # create empty file
-
-            with self.assertRaises(click.ClickException):
-                run_click_command(cci.project_init)
-
-    def test_project_init_dont_overwrite(self):
-        with temporary_dir():
-            # Gotta have a Repo
-            os.mkdir(".git")
-            Path(".git", "HEAD").write_text("ref: refs/heads/main")
-
-            os.mkdir("orgs")
-            orgs = "orgs/"
-            text = "Can't touch this"
-
-            path_list = [
-                Path("README.md"),
-                Path(".gitignore"),
-                Path(orgs + "dev.json"),
-                Path(orgs + "release.json"),
-            ]
-            for path in path_list:
-                path.write_text(text)
-
-            runtime = mock.Mock()
-            runtime.project_config.project = {"test": "test"}
-
-            run_click_command(cci.project_info, runtime=runtime)
-
-            # Project init must not overwrite project files or org defs
-            for path in path_list:
-                self.assertEqual(text, path.read_text())
-
-    @mock.patch("click.echo")
-    def test_project_info(self, echo):
-        runtime = mock.Mock()
-        runtime.project_config.project = {"test": "test"}
-
-        run_click_command(cci.project_info, runtime=runtime)
-
-        echo.assert_called_once_with("\x1b[1mtest:\x1b[0m test")
-
-    def test_project_info__outside_project(self):
-        runtime = mock.Mock()
-        runtime.project_config = None
-        runtime.project_config_error = NotInProject()
-        with temporary_dir():
-            with self.assertRaises(NotInProject):
-                run_click_command(cci.project_info, runtime=runtime)
-
-    @mock.patch("cumulusci.cli.cci.get_static_dependencies")
-    def test_project_dependencies(self, get_static_dependencies):
-        out = []
-        runtime = mock.Mock()
-        runtime.project_config.project__dependencies = [
-            {"namespace": "npe01", "version": "3.16"},
-            {"namespace": "npsp", "version": "3.193"},
-        ]
-        get_static_dependencies.return_value = [
-            PackageNamespaceVersionDependency(namespace="npe01", version="3.16"),
-            PackageNamespaceVersionDependency(namespace="npsp", version="3.193"),
-        ]
-
-        with mock.patch("click.echo", out.append):
-            run_click_command(
-                cci.project_dependencies,
-                runtime=runtime,
-                resolution_strategy="production",
-            )
-
-        self.assertEqual(
-            out,
-            [
-                str(
-                    PackageNamespaceVersionDependency(namespace="npe01", version="3.16")
-                ),
-                str(
-                    PackageNamespaceVersionDependency(namespace="npsp", version="3.193")
-                ),
-            ],
-        )
-
     @mock.patch("cumulusci.cli.cci.CliTable")
     def test_service_list(self, cli_tbl):
         runtime = mock.Mock()
@@ -943,9 +711,8 @@ Environment Info: Rossian / x68_46
         multi_cmd = cci.ConnectServiceCommand()
         runtime = mock.Mock()
         runtime.project_config.services = {"test": {}}
-        ctx = mock.Mock()
 
-        with mock.patch("cumulusci.cli.cci.RUNTIME", runtime):
+        with click.Context(multi_cmd, obj=runtime) as ctx:
             result = multi_cmd.list_commands(ctx)
         self.assertEqual(["test"], result)
 
@@ -954,23 +721,21 @@ Environment Info: Rossian / x68_46
         runtime = mock.Mock()
         runtime.project_config = None
         runtime.universal_config.services = {"test": {}}
-        ctx = mock.Mock()
 
-        with mock.patch("cumulusci.cli.cci.RUNTIME", runtime):
+        with click.Context(multi_cmd, obj=runtime) as ctx:
             result = multi_cmd.list_commands(ctx)
         self.assertEqual(["test"], result)
 
     def test_service_connect(self):
         multi_cmd = cci.ConnectServiceCommand()
-        ctx = mock.Mock()
         runtime = mock.MagicMock()
         runtime.project_config.services = {
             "test": {"attributes": {"attr": {"required": False}}}
         }
 
-        with mock.patch("cumulusci.cli.cci.RUNTIME", runtime):
+        with click.Context(multi_cmd, obj=runtime) as ctx:
             cmd = multi_cmd.get_command(ctx, "test")
-            run_click_command(cmd, project=True)
+            cmd.callback(ctx.obj, project=True)
 
         runtime.keychain.set_service.assert_called_once()
 
@@ -978,34 +743,31 @@ Environment Info: Rossian / x68_46
 
     def test_service_connect_global_keychain(self):
         multi_cmd = cci.ConnectServiceCommand()
-        ctx = mock.Mock()
         runtime = mock.MagicMock()
         runtime.project_config = None
         runtime.universal_config.services = {
             "test": {"attributes": {"attr": {"required": False}}}
         }
 
-        with mock.patch("cumulusci.cli.cci.RUNTIME", runtime):
+        with click.Context(multi_cmd, obj=runtime) as ctx:
             cmd = multi_cmd.get_command(ctx, "test")
-            run_click_command(cmd, project=True)
+            cmd.callback(ctx.obj, project=True)
 
-        runtime.keychain.set_service.assert_called_once()
+            runtime.keychain.set_service.assert_called_once()
 
-        run_click_command(cmd, project=False)
+            cmd.callback(ctx.obj, project=False)
 
     def test_service_connect_invalid_service(self):
         multi_cmd = cci.ConnectServiceCommand()
-        ctx = mock.Mock()
         runtime = mock.MagicMock()
         runtime.project_config.services = {}
 
-        with mock.patch("cumulusci.cli.cci.RUNTIME", runtime):
+        with click.Context(multi_cmd, obj=runtime) as ctx:
             with self.assertRaises(click.UsageError):
                 multi_cmd.get_command(ctx, "test")
 
     def test_service_connect_validator(self):
         multi_cmd = cci.ConnectServiceCommand()
-        ctx = mock.Mock()
         runtime = mock.MagicMock()
         runtime.project_config.services = {
             "test": {
@@ -1014,11 +776,10 @@ Environment Info: Rossian / x68_46
             }
         }
 
-        with mock.patch("cumulusci.cli.cci.RUNTIME", runtime):
+        with click.Context(multi_cmd, obj=runtime) as ctx:
             cmd = multi_cmd.get_command(ctx, "test")
-            with self.assertRaises(Exception) as cm:
-                run_click_command(cmd, project=True)
-            self.assertEqual("Validation failed", str(cm.exception))
+            with pytest.raises(Exception, match="Validation failed"):
+                cmd.callback(ctx.obj, project=True)
 
     @mock.patch("cumulusci.cli.cci.CliTable")
     def test_service_info(self, cli_tbl):
